refactor(carRacingProceeding, car): class Car 파일 분리
