+++ src/CarRacingGame/carRacingProceeding.js
@@ -1,24 +1,12 @@
-import { Console, Random } from '@woowacourse/mission-utils';
+import { Console } from '@woowacourse/mission-utils';
 import { inputCarNames, inputTryCount } from './input.js';
 import { printExecutionResult, printWinner } from './output.js';
 import { validateCarName, validateTryCount } from './validation.js';
+import { Car } from './car.js';
 
 export async function carRacingProceed() {
   const carNameList = await inputCarNames();
   validateCarName(carNameList);
-
-  class Car {
-    constructor(name, number) {
-      (this.name = name), (this.number = number);
-    }
-
-    move() {
-      let randomNum = Random.pickNumberInRange(0, 9);
-      if (randomNum >= 4) {
-        this.number += 1;
-      }
-    }
-  }
 
   const carList = [];
